Move type stats to separate file
