+++ templates/user.html
@@ -96,8 +96,10 @@
         </div>
         <h3>History</h3>
         <div id="history-stat-container" class="stat-container">
-          <div id="history-diagram"></div>
-          <div id="rate-diagram"></div>
+          <div id="gamma-history-diagram"></div>
+          <div id="gamma-rate-diagram"></div>
+          <div id="karma-history-diagram"></div>
+          <div id="karma-rate-diagram"></div>
         </div>
       </div>
     </div>
